add stack access validator
